requirements.txt
pandas>=1.5.0
google-cloud-bigquery>=3.10.0
pyarrow>=10.0.0
